Student-Performance-Pipeline/server/requirements.txt
flask
flask-cors
pandas
numpy
scikit-learn
psycopg2-binary
python-dotenv
seaborn
matplotlib
